shift2go_api/requirements.txt
alembic==1.4.3
Authlib==0.14.3
fastapi==0.63.0
celery==5.0.5
redis==3.5.3
flower==1.0.0
httpx==0.15.5
ipython==7.18.1
itsdangerous==1.1.0
Jinja2==2.11.3
psycopg2==2.8.6
pytest==6.1.0
pytest-cov==2.12.1
pytest-html==3.1.1
pytest-sugar==0.9.4
pytest-json-report==1.4.0
coverage==5.5.0
requests==2.24.0
SQLAlchemy==1.3.19
uvicorn==0.12.1
passlib==1.7.2
bcrypt==3.2.0
sqlalchemy-utils==0.36.8
python-multipart==0.0.5
pyjwt==1.7.1
phonenumbers==8.12.29
haversine==2.2.0
python-dotenv==0.19.0
firebase-admin==5.0.0